adding max number of files to process

+++ CMGTools/HtoZZ2l2nu/python/StandardSelections_cfi.py
@@ -45,8 +45,8 @@ BaseJetSelection = cms.PSet( source = cms.InputTag("selectedPatJets"),
                              )
 
 # base values for the dilepton selection ------------------------------------------
-BaseDileptonSelection = cms.PSet( minDileptonMass = cms.double(50),
-                                  maxDileptonMass = cms.double(7000)
+BaseDileptonSelection = cms.PSet( minDileptonMass = cms.double(71),
+                                  maxDileptonMass = cms.double(111)
                                   )
 
 # base values for met selection -----------------------------------------------------
